feat: add build script and devDependency for TypeScript 5.2.2
This commit adds a new "build" script to the package.json file, which runs the "pnpm -r build" command. Additionally, it adds "typescript": "5.2.2" as a devDependency in the package.json file and updates the pnpm-lock.yaml file accordingly. A new tsconfig.json file is also added with compiler options for targeting ES2023, using Node16 module resolution, and enabling source maps and inline sources in emitted JavaScript files. The strict flag is set to true for strict type-checking options.

Signed-off-by: Ardalan Amini <ardalanamini22@gmail.com>